requirements.txt
# Core dependencies
openai>=1.50.0  # Latest version for o1-mini support
ccxt>=4.4.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0

# Optional dependencies for web API
fastapi>=0.100.0
uvicorn>=0.20.0

# Development dependencies  
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
